change back to single swift executable example
